IsaacMCP/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isaac-mcp"
version = "0.1.0"
description = "MCP server bridging AI coding assistants with NVIDIA Isaac Sim"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "asyncssh>=2.14.0",
    "pyyaml>=6.0",
    "Pillow>=10.0",
    "PyJWT[crypto]>=2.10.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
ros2 = ["rclpy"]
drone-swarm = []
all = ["rclpy"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
isaac-mcp = "isaac_mcp.cli.main:cli"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["isaac_mcp*"]
